// models/NewsFeed.ts
/**
 * Winthrop Intelligence Internal API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.39.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { LinkCollection } from '../models/LinkCollection';
import { LinkCollection1 } from '../models/LinkCollection1';
import { HttpFile } from '../http/http';

export class NewsFeed {
    'id'?: number;
    'title'?: string;
    'body'?: string;
    'url'?: string;
    'sourceId'?: number;
    'sourceType'?: string;
    'links'?: LinkCollection;
    'tagsList'?: LinkCollection1;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string",
            "format": ""
        },
        {
            "name": "body",
            "baseName": "body",
            "type": "string",
            "format": ""
        },
        {
            "name": "url",
            "baseName": "url",
            "type": "string",
            "format": ""
        },
        {
            "name": "sourceId",
            "baseName": "source_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "sourceType",
            "baseName": "source_type",
            "type": "string",
            "format": ""
        },
        {
            "name": "links",
            "baseName": "links",
            "type": "LinkCollection",
            "format": ""
        },
        {
            "name": "tagsList",
            "baseName": "tags_list",
            "type": "LinkCollection1",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return NewsFeed.attributeTypeMap;
    }

    public constructor() {
    }
}